Labs/Airflow_Labs/Lab_3/requirements.txt
apache-airflow
pandas>=2.1.0
scikit-learn>=1.4.0
numpy>=1.26.0
jinja2>=3.1.3